start views reorg
